requirements.txt
fastapi
uvicorn[standard]
tflite-runtime
python-multipart
Pillow
